Remove file caching feature
The file caching feature, invoked by using the --reuse option, provides
little value to the user and adds complexity to the code. This commit
strips out all caching functionality.

The value that the --reuse option provided was that it allowed the user
to skip downloading a new OVAL file. A future commit will add an option
that allows the user to specify a local OVAL file in lieu of downloading
a fresh one.