Use UCT's JSON for scanning instead of oval and oscap
The scanning the system with oscap tool is slow. It also leads to some
extra complexity in manifest mode and in selecting an oval file, since a
different oval file is needed for each ubuntu codename. Some JSON
can be generated from the Ubuntu CVE Tracker (UCT) which is more concise
and contains exactly the information needed to determine what packages
are vulnerable.

By making a single call to `dpkg -l` and using the more focused JSON,
this commit improves performance by approximately 10x.