Use lazy loading to only query system when not in manifest mode
When in manifest mode, we don't care about the local system's codename,
ESM status, or installed packages. In fact, trying to determine these
things could cause an error if the local system isn't Ubuntu. In order
to allow users of other distributions to use CVEScan to analyze ubuntu
package manifest files, as well as skip unnecessary work, this commit
changes some of SysInfo's properties to be lazy-loaded.